+++ crates/backend-uzu/src/backends/common/kernel/mod.rs
@@ -1,6 +1,5 @@
 use crate::backends::common::Backend;
 
-pub mod attention_gemm;
 pub mod matmul;
 
 include!(concat!(env!("OUT_DIR"), "/traits.rs"));
@@ -9,7 +8,6 @@ pub trait Kernels: Sized {
     type Backend: Backend<Kernels = Self>;
 
     autogen_kernels!();
-    type AttentionGemmBlock: attention_gemm::AttentionGemmBackendBlock<Backend = Self::Backend>;
     type MatmulKernel: matmul::MatmulKernel<Backend = Self::Backend>;
 }
