+++ crates/backend-uzu/src/backends/metal/context.rs
@@ -52,6 +52,10 @@ pub struct MetalContext {
 }
 
 impl MetalContext {
+    pub fn supports_mxu(&self) -> bool {
+        self.device.supports_mxu()
+    }
+
     pub(crate) fn device_tier(&self) -> DeviceTier {
         self.device_tier
     }
@@ -257,5 +261,9 @@ impl Context for MetalContext {
 
     fn sparse_buffers_supported(&self) -> bool {
         self.device.supports_placement_sparse_resources()
+    }
+
+    fn supports_mxu(&self) -> bool {
+        self.supports_mxu()
     }
 }
