+++ crates/backend-uzu/src/backends/metal/mod.rs
@@ -12,4 +12,3 @@ mod sparse;
 pub use backend::Metal;
 pub use context::MetalContext;
 pub use kernel::matmul::GemmDispatchPath;
-pub(crate) use metal_extensions::DeviceExt;
