+++ crates/backend-uzu/unit/backends/common/kernel/matmul/quant_dispatch_test.rs
@@ -22,7 +22,7 @@ use crate::{
             },
         },
         cpu::Cpu,
-        metal::{DeviceExt, GemmDispatchPath, Metal, MetalContext},
+        metal::{GemmDispatchPath, Metal, MetalContext},
     },
     tests::{
         helpers::allocation_to_vec,
@@ -519,7 +519,7 @@ fn mxu_quant_parity_bf16(
     #[case] method: QuantizationMethod,
 ) {
     let context = MetalContext::new().expect("Metal context");
-    if !context.device.supports_mxu() {
+    if !context.supports_mxu() {
         return;
     }
     let input = QuantInput::<bf16>::new(m, k, n, gs, bits, method, 0);
