+++ crates/backend-uzu/unit/backends/common/kernel/matmul/quant_gemm_bench.rs
@@ -13,7 +13,7 @@ use crate::{
             gpu_types::QuantizationMethod,
             kernel::{Kernels, matmul::MatmulKernel},
         },
-        metal::{DeviceExt, GemmDispatchPath, Metal, MetalContext},
+        metal::{GemmDispatchPath, Metal, MetalContext},
     },
     tests::{
         matmul::{QuantBuffers, QuantInput, bench_quant_gemm_shapes, iter_encode_loop, quant_arguments},
@@ -29,7 +29,7 @@ fn bench_unified_quant_typed<T: ArrayElement + Float>(
     bits: u32,
     quant_method: QuantizationMethod,
 ) {
-    let supports_mxu = context.device.supports_mxu();
+    let supports_mxu = context.supports_mxu();
     let paths: &[(&str, GemmDispatchPath)] = if supports_mxu {
         &[("Simdgroup", GemmDispatchPath::Simdgroup), ("Mxu", GemmDispatchPath::Mxu)]
     } else {
